Move gradle utilities which aren't specific to the IDE task out of `EquoIde` and into `GradleCommon`.
